updated job scheduler and ThreatIntelFeedDataService
