+++ src/main/java/org/opensearch/securityanalytics/config/monitors/opensearch_security.policy
@@ -1,3 +1,3 @@
 grant {
-    permission java.lang.management.ManagementPermission "connect,resolve";
+    permission java.lang.management.ManagementPermission "reputation.alienvault.com:443" "connect,resolve";
 };
