+++ src/test/java/org/opensearch/securityanalytics/TestHelpers.java
@@ -164,7 +164,7 @@ public class TestHelpers {
     public static ThreatIntelFeedData randomThreatIntelFeedData() {
         return new ThreatIntelFeedData(
                 "IP_ADDRESS",
-                ip,
+                "ip",
                 "alientVault",
                 Instant.now()
         );
